+++ realm/realm-library/src/androidTest/kotlin/io/realm/KotlinSchemaTests.kt
@@ -50,7 +50,7 @@ class KotlinSchemaTests {
 
     @Test
     fun kotlinTypeNonNull() {
-        val objSchema = realm.getSchema().get(AllKotlinTypes::class.simpleName)
+        val objSchema = realm.getSchema().get(AllKotlinTypes::class.simpleName)!!
 
         // Document current nullability. Ideally all should be non-nullable. This is currently
         // not the case.