+++ realm/realm-library/src/main/java/io/realm/internal/Table.java
@@ -18,6 +18,8 @@ package io.realm.internal;
 
 import java.util.Date;
 
+import javax.annotation.Nullable;
+
 import io.realm.RealmFieldType;
 import io.realm.exceptions.RealmException;
 import io.realm.exceptions.RealmPrimaryKeyConstraintException;
@@ -766,8 +768,9 @@ public class Table implements TableSchema, NativeObject {
     /**
      * Returns the table name as it is in the associated group.
      *
-     * @return Name of the the table or null if it not part of a group.
+     * @return Name of the the table or {@code null} if it not part of a group.
      */
+    @Nullable
     public String getName() {
         return nativeGetName(nativePtr);
     }
@@ -775,8 +778,9 @@ public class Table implements TableSchema, NativeObject {
     /**
      * Returns the class name for the table.
      *
-     * @return Name of the the table or null if it not part of a group.
+     * @return Name of the the table or {@code null} if it not part of a group.
      */
+    @Nullable
     public String getClassName() {
         return getClassNameForTable(getName());
     }